Extract agent discovery, frontmatter parsing, rosters, and prompt assembly to lib/agents
Verbatim move of the agent area out of server.js:

- lib/agents/discovery.js: discoverAgents (with its roster cache),
  readNormalisedFile, titleCase, and the frontmatter parsers
  (extractFrontmatterText, parseAgentFrontmatter, parseCapabilities,
  parseRoutines, parsePrompts, parseSkills). Reads getWorkspace() at use
  time. The scheduler's routineState arrives via setRoutineState BY
  IDENTITY (mutated in place by the root) until the scheduler's own slice.

- lib/agents/prompt.js: buildTeamRoster, buildPeerRoster,
  extractSelfDescription, findDirectReportMatch,
  findOffRosterWorkspaceMatch, buildSystemPrompt. discoverSkills and
  detectCodexCached stay in the root (skill discovery is its own area;
  the codex probe cache is shared with the settings runtime probe) and
  arrive via wirePromptDeps; unwired deps throw at first use.

- The root's invalidateAgentCache() becomes the cascade caller: module
  agent cache first, then the root-owned skill/file-list/file-tree
  caches. The save_agent handler and its source pin are unchanged.

- _internal re-exports the module functions by identity; the whole
  existing characterization suite drives the moved code unchanged.

- New seam tests in test/unit/agents-lib.test.js: identity re-export,
  use-time workspace reads, routineState identity injection, and both
  prompt dep injections.
